+++ src/vue/common/AccountBalances.vue
@@ -1,27 +1,33 @@
 <template>
   <div class="account-balances">
     <template v-if="balancesAll && isLoadedStaking">
-      <p class="account-balances__total">
+      <p
+        v-tooltip="$fFullBalance(totalBalance)"
+        class="account-balances__total"
+      >
         {{ $fbalance(totalBalance) }}
       </p>
 
       <readonly-row
-        class="account-balances__row"
         v-if="!availableBalance.isZero()"
+        class="account-balances__row"
+        :tooltip="$fFullBalance(availableBalance)"
         :label="$t('common.account-balances.transferrable-lbl')"
         :value="$fbalance(availableBalance)"
       />
 
       <readonly-row
-        class="account-balances__row"
         v-if="!balancesAll.lockedBalance.isZero()"
+        class="account-balances__row"
+        :tooltip="$fFullBalance(balancesAll.lockedBalance)"
         :label="$t('common.account-balances.locked-lbl')"
         :value="$fbalance(balancesAll.lockedBalance)"
       />
 
       <readonly-row
-        class="account-balances__row"
         v-if="+stakingBalance.balance && !isLoadStakingFailed"
+        class="account-balances__row"
+        :tooltip="$fFullBalance(stakingBalance.balance)"
         :label="$t('common.account-balances.staked-lbl')"
         :value="$fbalance(stakingBalance.balance)"
       />
@@ -111,6 +117,7 @@ export default {
 .account-balances__total {
   font-size: 2rem;
   color: $col-app-accent;
+  width: max-content;
 }
 
 .account-balances__row {