+++ src/vue/pages/explorer/tabs/block-info/BlockInfoSummary.vue
@@ -3,12 +3,14 @@
     <div class="block-info-summary__content">
       <info-block
         class="block-info-summary__info"
+        v-tooltip="$fFullBalance(deposits)"
         :title="$t('explorer-page.block-info-summary.header-deposits')"
         :value="$fbalance(deposits)"
       />
 
       <info-block
         class="block-info-summary__info"
+        v-tooltip="$fFullBalance(transfers)"
         :title="$t('explorer-page.block-info-summary.header-transfers')"
         :value="$fbalance(transfers)"
       />